Less things in the Types module
