+++ marconi-cardano-core/src/Marconi/Cardano/Core/Types.hs
@@ -1,9 +1,8 @@
 {-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE ExistentialQuantification #-}
 {-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE PatternSynonyms #-}
-{-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE TypeFamilies #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
 
 -- | This module provides several type aliases and utility functions to deal with them.
 module Marconi.Cardano.Core.Types (
@@ -23,9 +22,6 @@ module Marconi.Cardano.Core.Types (
   TipAndBlock (..),
 
   -- * Aliases for the current Cardano era
-  CurrentEra,
-  pattern AsCurrentEra,
-  pattern CurrentEra,
   TxOut,
 
   -- * Aliases to ease concept mapping between plutus types and cardano types
@@ -35,28 +31,13 @@ module Marconi.Cardano.Core.Types (
 
   -- * Reexport from cardano-api-extended
   BlockEvent (..),
-
-  -- * Database file names
-  utxoDbName,
-  addressDatumDbName,
-  scriptTxDbName,
-  epochStateDbName,
-  mintBurnDbName,
   SecurityParam (SecurityParam),
   TxIndexInBlock (TxIndexInBlock),
-
-  -- * Block range type for specifying a sub-chain
-  BlockRange,
-  mkBlockRange,
-  blockRangeFst,
-  blockRangeSnd,
-  isInBlockRange,
 ) where
 
 import Cardano.Api qualified as C
 import Cardano.Api.Extended.Streaming (BlockEvent (BlockEvent, blockInMode, blockTime, epochNo))
 import Cardano.BM.Data.Trace (Trace)
-import Control.Lens.TH qualified as Lens
 import Data.Aeson (FromJSON, ToJSON)
 import Data.Aeson qualified as Aeson
 import Data.Set.NonEmpty (NESet)
@@ -66,6 +47,7 @@ import Database.SQLite.Simple.ToField qualified as SQL
 import GHC.Generics (Generic)
 import Marconi.Cardano.Core.Extract.WithDistance (WithDistance)
 import Marconi.Core qualified as Core
+import Ouroboros.Consensus.Config (SecurityParam (SecurityParam))
 import Prettyprinter (Doc)
 
 -- Point type instances for types of this module
@@ -102,17 +84,8 @@ data TipAndBlock = TipAndBlock C.ChainTip (Maybe (WithDistance BlockEvent))
 
 type instance Core.Point TipAndBlock = C.ChainPoint
 
--- | An alias for the current era, to ease the transition from one era to the next one
-type CurrentEra = C.BabbageEra
-
-pattern CurrentEra :: C.CardanoEra CurrentEra
-pattern CurrentEra = C.BabbageEra
-
-pattern AsCurrentEra :: C.AsType CurrentEra
-pattern AsCurrentEra = C.AsBabbageEra
-
 -- | A Cardano TxOut of the current Era
-type TxOut = C.TxOut C.CtxTx CurrentEra
+type TxOut = C.TxOut C.CtxTx C.ConwayEra
 
 {- | A reference to a transaction output. This is a
  pair of a transaction reference, and an index indicating which of the outputs
@@ -123,8 +96,13 @@ type TxOutRef = C.TxIn
 txOutRef :: C.TxId -> C.TxIx -> C.TxIn
 txOutRef = C.TxIn
 
-newtype SecurityParam = SecurityParam Word64
-  deriving newtype (Eq, Ord, Bounded, Enum, Real, Num, Read, Integral, Show)
+deriving newtype instance Ord SecurityParam
+deriving newtype instance Enum SecurityParam
+deriving newtype instance Real SecurityParam
+deriving newtype instance Num SecurityParam
+deriving newtype instance Integral SecurityParam
+deriving newtype instance SQL.ToField SecurityParam
+deriving newtype instance SQL.FromField SecurityParam
 
 newtype TxIndexInBlock = TxIndexInBlock Word64
   deriving newtype
@@ -145,42 +123,3 @@ newtype TxIndexInBlock = TxIndexInBlock Word64
 
 -- | An existential type representing a transaction with @C.'TxBody' era@ for any Cardano era.
 data AnyTxBody = forall era. (C.IsCardanoEra era) => AnyTxBody C.BlockNo TxIndexInBlock (C.TxBody era)
-
--- * Database file names
-
-utxoDbName :: FilePath
-utxoDbName = "utxo.db"
-
-addressDatumDbName :: FilePath
-addressDatumDbName = "addressdatum.db"
-
-scriptTxDbName :: FilePath
-scriptTxDbName = "scripttx.db"
-
-epochStateDbName :: FilePath
-epochStateDbName = "epochstate.db"
-
-mintBurnDbName :: FilePath
-mintBurnDbName = "mintburn.db"
-
-data BlockRange = BlockRange
-  { _blockRangeFst :: !Word64
-  , _blockRangeSnd :: !Word64
-  }
-  deriving stock (Show, Generic)
-  deriving anyclass (FromJSON, ToJSON)
-
-Lens.makeLenses ''BlockRange
-
-mkBlockRange :: Word64 -> Word64 -> Either String BlockRange
-mkBlockRange x y
-  | x <= 0 || y <= 0 = Left "Expected positive arguments in block range."
-  | x <= y = Right $ BlockRange x y
-  | otherwise =
-      Left $
-        "Expected left hand side of the block range "
-          <> "to be smaller than or equal to the right hand side."
-
-isInBlockRange :: C.BlockNo -> BlockRange -> Bool
-isInBlockRange (C.BlockNo bNo) (BlockRange left right) =
-  left <= bNo && bNo <= right
